players updates
Added a 100 person random array in a component. added a modal to display the rollcall. added a rollcall component for the modal. player can be marked present now and stored in state. added a state with just the present players that can transfer to manage players (small issue with real time here)